// src/app/models/cliente.ts
// export interface Perfil {
//     username?: string;
//     email?: string;
//     password?: string;
//     lastname?: string;
//     phone?: string;
//     address?: string;
//     activo?: boolean;
//     photoUrl?: string;
// }

export interface UserInterface {
    // perfil: Perfil;
    username?: string;
    email?: string;
    password?: string;
    lastname?: string;
    phone?: string;
    address?: string;
    activo?: boolean;
    photoUrl?: string;
    tipo?: number;
    uid?: string;
    uidSucursal?: string;

}